Version 0.0.3
Incompatible change:- |operator#user#define()|, |operator#user#define_ex_command()|: Change the meaning of the first argument:  Old versions:	  The argument is {lhs} to execute the operator to be defined.  New version:	  The argument is {name} of the operator to be defined.	  {lhs} to execute the operator is "<Plug>(operator-{name})".  See also |operator-user-bugs-C| for the details.